feat: introduce 'skip' config key
