+++ Cargo.toml
@@ -223,3 +223,6 @@ hyper = "1.0"
 reqwest = { version = "0.12", default-features = false }
 tower = "0.4"
 tower-http = "0.5"
+
+[patch.crates-io]
+foundry-compilers = { path = "../compilers" }
