+++ crates/common/Cargo.toml
@@ -39,8 +39,6 @@ clap = { version = "4", features = ["derive", "env", "unicode", "wrap_help"] }
 comfy-table = "7"
 dunce = "1"
 eyre.workspace = true
-glob = "0.3"
-globset = "0.4"
 once_cell = "1"
 reqwest.workspace = true
 semver = "1"
