+++ crates/common/src/lib.rs
@@ -17,7 +17,6 @@ pub mod errors;
 pub mod evm;
 pub mod fmt;
 pub mod fs;
-pub mod glob;
 pub mod provider;
 pub mod retry;
 pub mod selectors;
