+++ crates/config/Cargo.toml
@@ -25,6 +25,7 @@ dunce = "1"
 eyre.workspace = true
 figment = { version = "0.10", features = ["toml", "env"] }
 globset = "0.4"
+glob = "0.3"
 Inflector = "0.11"
 number_prefix = "0.4"
 once_cell = "1"
