+++ crates/config/src/filter.rs
@@ -1,0 +1,203 @@
+//! Helpers for constructing and using [FileFilter]s.
+
+use core::fmt;
+use foundry_compilers::FileFilter;
+use std::{
+    convert::Infallible,
+    path::{Path, PathBuf},
+    str::FromStr,
+};
+
+/// Expand globs with a root path.
+pub fn expand_globs(
+    root: &Path,
+    patterns: impl IntoIterator<Item = impl AsRef<str>>,
+) -> eyre::Result<Vec<PathBuf>> {
+    let mut expanded = Vec::new();
+    for pattern in patterns {
+        for paths in glob::glob(&root.join(pattern.as_ref()).display().to_string())? {
+            expanded.push(paths?);
+        }
+    }
+    Ok(expanded)
+}
+
+/// A `globset::Glob` that creates its `globset::GlobMatcher` when its created, so it doesn't need
+/// to be compiled when the filter functions `TestFilter` functions are called.
+#[derive(Clone, Debug)]
+pub struct GlobMatcher {
+    /// The compiled glob
+    pub matcher: globset::GlobMatcher,
+}
+
+impl GlobMatcher {
+    /// Creates a new `GlobMatcher` from a `globset::Glob`.
+    pub fn new(glob: globset::Glob) -> Self {
+        Self { matcher: glob.compile_matcher() }
+    }
+
+    /// Tests whether the given path matches this pattern or not.
+    ///
+    /// The glob `./test/*` won't match absolute paths like `test/Contract.sol`, which is common
+    /// format here, so we also handle this case here
+    pub fn is_match(&self, path: &Path) -> bool {
+        if self.matcher.is_match(path) {
+            return true;
+        }
+
+        if !path.starts_with("./") && self.as_str().starts_with("./") {
+            return self.matcher.is_match(format!("./{}", path.display()));
+        }
+
+        false
+    }
+
+    /// Returns the `globset::Glob`.
+    pub fn glob(&self) -> &globset::Glob {
+        self.matcher.glob()
+    }
+
+    /// Returns the `Glob` string used to compile this matcher.
+    pub fn as_str(&self) -> &str {
+        self.glob().glob()
+    }
+}
+
+impl fmt::Display for GlobMatcher {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.glob().fmt(f)
+    }
+}
+
+impl FromStr for GlobMatcher {
+    type Err = globset::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        s.parse::<globset::Glob>().map(Self::new)
+    }
+}
+
+impl From<globset::Glob> for GlobMatcher {
+    fn from(glob: globset::Glob) -> Self {
+        Self::new(glob)
+    }
+}
+
+/// Bundles multiple `SkipBuildFilter` into a single `FileFilter`
+#[derive(Clone, Debug)]
+pub struct SkipBuildFilters {
+    /// All provided filters.
+    pub matchers: Vec<GlobMatcher>,
+    /// Root of the project.
+    pub project_root: PathBuf,
+}
+
+impl FileFilter for SkipBuildFilters {
+    /// Only returns a match if _no_  exclusion filter matches
+    fn is_match(&self, file: &Path) -> bool {
+        self.matchers.iter().all(|matcher| {
+            if !is_match_exclude(matcher, file) {
+                false
+            } else {
+                file.strip_prefix(&self.project_root)
+                    .map_or(true, |stripped| is_match_exclude(matcher, stripped))
+            }
+        })
+    }
+}
+
+impl SkipBuildFilters {
+    /// Creates a new `SkipBuildFilters` from multiple `SkipBuildFilter`.
+    pub fn new<G: Into<GlobMatcher>>(
+        filters: impl IntoIterator<Item = G>,
+        project_root: PathBuf,
+    ) -> Self {
+        let matchers = filters.into_iter().map(|m| m.into()).collect();
+        Self { matchers, project_root }
+    }
+}
+
+/// A filter that excludes matching contracts from the build
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum SkipBuildFilter {
+    /// Exclude all `.t.sol` contracts
+    Tests,
+    /// Exclude all `.s.sol` contracts
+    Scripts,
+    /// Exclude if the file matches
+    Custom(String),
+}
+
+impl SkipBuildFilter {
+    fn new(s: &str) -> Self {
+        match s {
+            "test" | "tests" => SkipBuildFilter::Tests,
+            "script" | "scripts" => SkipBuildFilter::Scripts,
+            s => SkipBuildFilter::Custom(s.to_string()),
+        }
+    }
+
+    /// Returns the pattern to match against a file
+    pub fn file_pattern(&self) -> &str {
+        match self {
+            SkipBuildFilter::Tests => ".t.sol",
+            SkipBuildFilter::Scripts => ".s.sol",
+            SkipBuildFilter::Custom(s) => s.as_str(),
+        }
+    }
+}
+
+impl FromStr for SkipBuildFilter {
+    type Err = Infallible;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(Self::new(s))
+    }
+}
+
+/// Matches file only if the filter does not apply.
+///
+/// This returns the inverse of `file.name.contains(pattern) || matcher.is_match(file)`.
+fn is_match_exclude(matcher: &GlobMatcher, path: &Path) -> bool {
+    fn is_match(matcher: &GlobMatcher, path: &Path) -> Option<bool> {
+        let file_name = path.file_name()?.to_str()?;
+        Some(file_name.contains(matcher.as_str()) || matcher.is_match(path))
+    }
+
+    !is_match(matcher, path).unwrap_or_default()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_build_filter() {
+        let tests = GlobMatcher::from_str(SkipBuildFilter::Tests.file_pattern()).unwrap();
+        let scripts = GlobMatcher::from_str(SkipBuildFilter::Scripts.file_pattern()).unwrap();
+        let custom = |s| GlobMatcher::from_str(s).unwrap();
+
+        let file = Path::new("A.t.sol");
+        assert!(!is_match_exclude(&tests, file));
+        assert!(is_match_exclude(&scripts, file));
+        assert!(!is_match_exclude(&custom("A.t"), file));
+
+        let file = Path::new("A.s.sol");
+        assert!(is_match_exclude(&tests, file));
+        assert!(!is_match_exclude(&scripts, file));
+        assert!(!is_match_exclude(&custom("A.s"), file));
+
+        let file = Path::new("/home/test/Foo.sol");
+        assert!(!is_match_exclude(&custom("*/test/**"), file));
+
+        let file = Path::new("/home/script/Contract.sol");
+        assert!(!is_match_exclude(&custom("*/script/**"), file));
+    }
+
+    #[test]
+    fn can_match_glob_paths() {
+        let matcher: GlobMatcher = "./test/*".parse().unwrap();
+        assert!(matcher.is_match(Path::new("test/Contract.sol")));
+        assert!(matcher.is_match(Path::new("./test/Contract.sol")));
+    }
+}
