+++ crates/doc/Cargo.toml
@@ -12,7 +12,6 @@ repository.workspace = true
 
 [dependencies]
 forge-fmt.workspace = true
-foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
