+++ crates/doc/src/builder.rs
@@ -3,9 +3,8 @@ use crate::{
     ParseSource, Parser, Preprocessor,
 };
 use forge_fmt::{FormatterConfig, Visitable};
-use foundry_common::glob::expand_globs;
 use foundry_compilers::{utils::source_files_iter, SOLC_EXTENSIONS};
-use foundry_config::DocConfig;
+use foundry_config::{filter::expand_globs, DocConfig};
 use itertools::Itertools;
 use mdbook::MDBook;
 use rayon::prelude::*;
