+++ crates/forge/bin/cmd/bind.rs
@@ -6,6 +6,7 @@ use eyre::{Result, WrapErr};
 use foundry_cli::{opts::CoreBuildArgs, utils::LoadConfig};
 use foundry_common::{compile::ProjectCompiler, fs::json_files};
 use foundry_config::impl_figment_convert;
+use regex::Regex;
 use std::{
     fs,
     path::{Path, PathBuf},
@@ -31,10 +32,6 @@ pub struct BindArgs {
     /// Create bindings only for contracts whose names match the specified filter(s)
     #[arg(long)]
     pub select: Vec<regex::Regex>,
-
-    /// Create bindings only for contracts whose names do not match the specified filter(s)
-    #[arg(long, conflicts_with = "select")]
-    pub skip: Vec<regex::Regex>,
 
     /// Explicitly generate bindings for all contracts
     ///
@@ -133,18 +130,25 @@ impl BindArgs {
     }
 
     /// Returns the filter to use for `MultiAbigen`
-    fn get_filter(&self) -> ContractFilter {
+    fn get_filter(&self) -> Result<ContractFilter> {
         if self.select_all {
-            return ContractFilter::All
+            return Ok(ContractFilter::All)
         }
         if !self.select.is_empty() {
-            return SelectContracts::default().extend_regex(self.select.clone()).into()
+            return Ok(SelectContracts::default().extend_regex(self.select.clone()).into())
         }
-        if !self.skip.is_empty() {
-            return ExcludeContracts::default().extend_regex(self.skip.clone()).into()
+        if let Some(skip) = self.build_args.skip.as_ref().filter(|s| !s.is_empty()) {
+            return Ok(ExcludeContracts::default()
+                .extend_regex(
+                    skip.clone()
+                        .into_iter()
+                        .map(|s| Regex::new(s.file_pattern()))
+                        .collect::<Result<Vec<_>, _>>()?,
+                )
+                .into())
         }
         // This excludes all Test/Script and forge-std contracts
-        ExcludeContracts::default()
+        Ok(ExcludeContracts::default()
             .extend_pattern([
                 ".*Test.*",
                 ".*Script",
@@ -155,13 +159,13 @@ impl BindArgs {
                 "[Vv]m.*",
             ])
             .extend_names(["IMulticall3"])
-            .into()
+            .into())
     }
 
     /// Returns an iterator over the JSON files and the contract name in the `artifacts` directory.
-    fn get_json_files(&self, artifacts: &Path) -> impl Iterator<Item = (String, PathBuf)> {
-        let filter = self.get_filter();
-        json_files(artifacts)
+    fn get_json_files(&self, artifacts: &Path) -> Result<impl Iterator<Item = (String, PathBuf)>> {
+        let filter = self.get_filter()?;
+        Ok(json_files(artifacts)
             .filter_map(|path| {
                 // Ignore the build info JSON.
                 if path.to_str()?.contains("/build-info/") {
@@ -181,13 +185,13 @@ impl BindArgs {
 
                 Some((name, path))
             })
-            .filter(move |(name, _path)| filter.is_match(name))
+            .filter(move |(name, _path)| filter.is_match(name)))
     }
 
     /// Instantiate the multi-abigen
     fn get_multi(&self, artifacts: &Path) -> Result<MultiAbigen> {
         let abigens = self
-            .get_json_files(artifacts)
+            .get_json_files(artifacts)?
             .map(|(name, path)| {
                 trace!(?path, "parsing Abigen from file");
                 let abi = Abigen::new(name, path.to_str().unwrap())
