+++ crates/forge/tests/cli/build.rs
@@ -1,4 +1,6 @@
+use foundry_config::Config;
 use foundry_test_utils::{forgetest, util::OutputExt};
+use globset::Glob;
 use std::path::PathBuf;
 
 // tests that json is printed when --json is passed
@@ -39,4 +41,31 @@ forgetest_init!(build_sizes_no_forge_std, |prj, cmd| {
     assert!(!stdout.contains("console"), "\n{stdout}");
     assert!(!stdout.contains("std"), "\n{stdout}");
     assert!(stdout.contains("Counter"), "\n{stdout}");
+});
+
+// tests that skip key in config can be used to skip non-compilable contract
+forgetest_init!(test_can_skip_contract, |prj, cmd| {
+    prj.add_source(
+        "InvalidContract",
+        r"
+contract InvalidContract {
+    some_invalid_syntax
+}
+",
+    )
+    .unwrap();
+
+    prj.add_source(
+        "ValidContract",
+        r"
+contract ValidContract {}
+",
+    )
+    .unwrap();
+
+    let config =
+        Config { skip: vec![Glob::new("src/InvalidContract.sol").unwrap()], ..Default::default() };
+    prj.write_config(config);
+
+    cmd.args(["build"]).assert_success();
 });
